CHANGE INT TO BIGINT

+++ sql_scripts/create/reservation.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS reservation;
 CREATE TABLE reservation (
-    id_reservation INT AUTO_INCREMENT PRIMARY KEY,
-    id_user INT,
+    id_reservation BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id_user BIGINT,
     number_room INT,
-    reference_promo INT,
+    reference_promo BIGINT,
     start_date_reservation DATE,
     end_date_reservation DATE,
     nbr_person_reservation INT,
