+++ sql_scripts/create/user_company.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS user_company;
 CREATE TABLE user_company(
-    id_user_company INT AUTO_INCREMENT PRIMARY KEY,
-    id_company INT,
-    id_user INT,
+    id_user_company BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id_company BIGINT,
+    id_user BIGINT,
     FOREIGN KEY (id_company) REFERENCES company(id_company),
     FOREIGN KEY (id_user) REFERENCES user(id_user)
 );
